+++ lesson1/my-app/src/components/index.js
@@ -1,3 +1,4 @@
 export * from './Message'
 export * from './Button'
-export * from './ChatsList'
+export * from './Chats'
+export * from './Header'
