+++ lesson1/my-app/src/routes/ChatRoute/index.js
@@ -1,0 +1,10 @@
+
+
+export const Chat = () => {
+
+    return (
+        <div>
+            <h1>Chat</h1>
+        </div>
+    )
+}
