+++ lesson1/my-app/src/routes/HomeRoute/index.js
@@ -1,0 +1,9 @@
+import { React } from 'react';
+export const Home = () => {
+
+
+    return (
+        <h1>HomePage</h1>
+    )
+}
+
