+++ lesson1/my-app/src/routes/Page404/index.js
@@ -1,0 +1,11 @@
+
+
+export const Page404 = () => {
+
+
+    return (
+        <div>
+            <h1>404</h1>
+        </div>
+    )
+}
