+++ lesson1/my-app/src/routes/Profile/index.js
@@ -1,0 +1,11 @@
+
+
+export const Profile = () => {
+
+
+    return (
+        <div>
+            <h1>Страница профиля</h1>
+        </div>
+    )
+}
